+++ CHANGELOG.md
@@ -23,3 +23,5 @@
 - Developed a Contact Inquiries module with view and delete capabilities.
 - Introduced load more functionality to improve performance by avoiding large data fetches at once.
 - Implemented property detail page
+- Implemented RETS Script
+- Implemented listing detail page